Change Build Driver To 'docker'
The default build driver for the used buildx action is `docker-container`.
This means the image is build within a buildkit container with no access
to images loaded into docker.
This is a problem because we need to import the base images for
parser-sdk and hook-sdk to build the images for parsers and hooks
without using a container registry caching these base images.

To fix this I changed the build driver to `docker` to build the image
with the docker daemon.

See:
https://github.com/docker/setup-buildx-action
github.com/docker/buildx/blob/master/docs/reference/buildx_create.md#driver

Signed-off-by: Yannik Fuhrmeister <yannik.fuhrmeister@iteratec.com>