Cache Operator and Lurker Image
I removed the docker build and push action and replaced it with the
build provided by the make file.
I also export the generated image and upload it as an artifact.
For the integration tests I ensured that the image artifacts are
downloaded and loaded into kind.

I changed the pullPolicy for operator and lurker that it won't try to
load their images from dockerhub and use the local images instead.

Signed-off-by: Yannik Fuhrmeister <yannik.fuhrmeister@iteratec.com>